google.apis/requirements.txt
google-adk>=0.1.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
litellm>=1.0.0
